LayoutTests/webgpu/buffers.html

<!DOCTYPE html><!-- webkit-test-runner [ experimental:WebGPUEnabled=true ] -->
<meta charset=utf-8>
<title>Get the WebGPUDevice, create a WebGPUBuffer, and write vertex data to it.</title>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/js-test-pre.js"></script>
<script src="js/basic-webgpu-functions.js"></script>
<script>
function createBuffer() {
    const buffer = defaultDevice.createBuffer({ size: 16, usage: GPUBufferUsage.MAP_WRITE });
    assert_true(buffer instanceof WebGPUBuffer, "createBuffer returned a WebGPUBuffer");
    
    let arrayBuffer = buffer.mapping;
    let floatArray = new Float32Array(arrayBuffer);
    assert_equals(floatArray.length, 4);

    floatArray[0] = -1;
    floatArray[1] = 1;
    floatArray[2] = 0;
    floatArray[3] = 1;
}

promise_test(async t => {
    const canvas = document.createElement("canvas");
    await setUpContexts(canvas);
    createBuffer();
}, "createBuffer() on WebGPUDevice.");

</script>
</body>
