LayoutTests/webgpu/buffers-expected.txt

PASS [object WebGPU] is defined.

PASS createBuffer() on WebGPUDevice. 
